[SERVER] organize package
