+++ rts-server/src/controllers/mod.rs
@@ -1,0 +1,3 @@
+pub mod ai_controller;
+pub mod leader_board_controller;
+pub mod user_controller;
